requirements.txt
streamlit
langchain
sentence-transformers
chromadb
pypdf
python-docx
ollama